# docs/swagger.yaml
basePath: /
definitions:
...
  internal_user.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
  internal_user.UpdateProfileRequest:
    properties:
      about:
        type: string
      position:
        type: string
      skills:
        items:
          type: string
        type: array
    type: object
  internal_user.UpdateWorkStatusRequest:
    properties:
      work_status:
        type: string
    required:
    - work_status
    type: object
  internal_user.UserResponse:
    properties:
      data:
        properties:
          about:
            type: string
          created_at:
            type: string
          email:
            type: string
          firstname:
            type: string
          id:
            type: string
          lastname:
            type: string
          position:
            type: string
          role:
            type: string
          skills:
            items:
              type: string
            type: array
          status:
            type: string
          work_status:
            type: string
        type: object
    type: object
  internal_user.WorkStatusHistoryEntry:
    properties:
      changed_at:
        type: string
      changed_by:
        type: string
      id:
        type: string
      user_id:
        type: string
      work_status:
        type: string
    type: object
  internal_user.WorkStatusHistoryResponse:
    properties:
      data:
        properties:
          history:
            items:
              $ref: '#/definitions/internal_user.WorkStatusHistoryEntry'
            type: array
        type: object
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend API для работы с командой над проектом.
  title: timeManager API
  version: "1.0"
paths:
...
  /api/v1/users/me:
    get:
      consumes:
      - application/json
      description: Возвращает данные авторизованного пользователя. Требует Bearer
        token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_user.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_user.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Текущий пользователь
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: Обновляет поля about (о себе), position (должность), skills (скиллы).
        Скиллы создаются в БД при первом добавлении.
      parameters:
      - description: Данные профиля
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_user.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_user.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_user.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_user.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить профиль текущего пользователя
      tags:
      - user
  /api/v1/users/me/avatar:
    put:
      consumes:
      - multipart/form-data
      description: multipart/form-data с полем avatar (jpeg/png/webp/gif, max 5MB).
      parameters:
      - description: Файл изображения
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_user.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_user.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_user.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Установить аватарку текущего пользователя
      tags:
      - user
  /api/v1/users/me/work-status:
    put:
      consumes:
      - application/json
      description: 'Обновляет рабочий статус: working, resting, lunch, vacation, sick_leave,
        business_trip. Запись добавляется в историю.'
      parameters:
      - description: Рабочий статус
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_user.UpdateWorkStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_user.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_user.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_user.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить рабочий статус текущего пользователя
      tags:
      - user
  /api/v1/users/me/work-status/history:
    get:
      consumes:
      - application/json
      description: Возвращает историю изменений рабочего статуса текущего пользователя.
      parameters:
      - description: Лимит записей (по умолчанию 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.WorkStatusHistoryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_user.ErrorResponse'
      security:
      - BearerAuth: []
      summary: История изменений рабочего статуса
      tags:
      - user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"